+++ src/super_gradients/recipes/training_hyperparams/coco2017_yolov5_train_params.yaml
@@ -26,7 +26,7 @@ mixed_precision: False
 valid_metrics_list:
   - _target_: super_gradients.training.metrics.DetectionMetrics
     post_prediction_callback:
-      _target_: super_gradients.training.models.detection_models.yolov5.YoloV5PostPredictionCallback
+      _target_: super_gradients.training.models.detection_models.yolov5_base.YoloV5PostPredictionCallback
     num_cls: 80
 
 metric_to_watch: 'mAP@0.50:0.95'
